requirements.txt
numpy
opencv-python
tensorflow
scipy
Pillow
django
tqdm